Pass through alter flag
